Рефакторим utils модуль
